script_modal refactored
